generator/requirements.txt
kafka-python-ng==2.2.3
Faker==33.0.0
psycopg2-binary==2.9.10
